Move the UI files to the API package
