+++ gni-system/src/main/java/api/ApiService.java
@@ -1,5 +1,6 @@
 package api;
 
+import api.test.Test;
 import com.google.gson.Gson;
 import com.thetransactioncompany.jsonrpc2.JSONRPC2Error;
 import com.thetransactioncompany.jsonrpc2.JSONRPC2ParseException;
