+++ gni-system/src/main/java/api/IncorrectInputException.java
@@ -1,4 +1,4 @@
-package ui;
+package api;
 
 /**
  * @author Noel
