+++ gni-system/src/main/java/api/UIService.java
@@ -1,4 +1,4 @@
-package ui;
+package api;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
