+++ gni-system/src/main/java/util/BootSystem.java
@@ -7,7 +7,7 @@ import pin.PinServiceMain;
 import systeminformation.SystemInformationServiceMain;
 import transactionin.TransactionReceiveServiceMain;
 import transactionout.TransactionDispatchServiceMain;
-import ui.UIServiceMain;
+import api.UIServiceMain;
 import users.UsersServiceMain;
 
 /**
