+++ gni-system/src/test/java/system/SystemTest.java
@@ -1,24 +1,12 @@
 package system;
 
-import api.ApiServiceMain;
-import authentication.AuthenticationServiceMain;
 import com.google.gson.Gson;
 import databeans.*;
-import io.advantageous.boon.core.Sys;
 import io.advantageous.qbit.http.client.HttpClient;
-import ledger.LedgerServiceMain;
-import pin.PinServiceMain;
-import transactionin.TransactionReceiveServiceMain;
-import transactionout.TransactionDispatchServiceMain;
-import ui.UIServiceMain;
-import users.UsersServiceMain;
-import util.BootSystem;
 import util.JSONParser;
-import util.TableCreator;
 
 import java.util.ArrayList;
 
-import static io.advantageous.qbit.http.client.HttpClientBuilder.httpClientBuilder;
 import static java.net.HttpURLConnection.HTTP_OK;
 
 /**
